+++ test/libsolidity/semanticTests/shifts/shift_overflow.sol
@@ -3,7 +3,7 @@ contract C {
         return x << y;
     }
 
-    function leftS(int8 x, int8 y) public returns (int8) {
+    function leftS(int8 x, uint8 y) public returns (int8) {
         return x << y;
     }
 }
@@ -14,5 +14,5 @@ contract C {
 // leftU(uint8,uint8): 255, 8 -> 0
 // leftU(uint8,uint8): 255, 1 -> 254
 // leftU(uint8,uint8): 255, 0 -> 255
-// leftS(int8,int8): 1, 7 -> -128 # Result is -128 and output is sign-extended, not zero-padded. #
-// leftS(int8,int8): 1, 6 -> 64
+// leftS(int8,uint8): 1, 7 -> -128 # Result is -128 and output is sign-extended, not zero-padded. #
+// leftS(int8,uint8): 1, 6 -> 64